+++ src/dependencies.rs
@@ -90,6 +90,14 @@ fn build_dependencies_inner(
         build.arg(format!("--target={target}"));
     }
 
+    if let Some(packages) = &info.build_std {
+        if packages.is_empty() {
+            build.arg("-Zbuild-std");
+        } else {
+            build.arg(format!("-Zbuild-std={packages}"));
+        }
+    }
+
     // Reusable closure for setting up the environment both for artifact generation and `cargo_metadata`
     let set_locking = |cmd: &mut Command| {
         if let OutputConflictHandling::Error = config.output_conflict_handling {
@@ -166,7 +174,7 @@ fn build_dependencies_inner(
                     .target
                     .crate_types
                     .iter()
-                    .all(|ctype| !matches!(ctype.as_str(), "proc-macro" | "lib"))
+                    .all(|ctype| !matches!(ctype.as_str(), "proc-macro" | "lib" | "rlib"))
                 {
                     continue;
                 }
@@ -174,7 +182,10 @@ fn build_dependencies_inner(
                     import_paths.insert(filename.parent().unwrap().into());
                 }
                 let package_id = artifact.package_id;
-                if let Some(prev) = artifacts.insert(package_id.clone(), Ok(artifact.filenames)) {
+                if let Some(prev) = artifacts.insert(
+                    package_id.clone(),
+                    Ok((artifact.target.name, artifact.filenames)),
+                ) {
                     artifacts.insert(
                         package_id.clone(),
                         Err(format!(
@@ -252,7 +263,7 @@ fn build_dependencies_inner(
             .unwrap();
 
         // Then go over all of its dependencies
-        let dependencies = root
+        let mut dependencies = root
             .dependencies
             .iter()
             .filter(|dep| matches!(dep.kind, DependencyKind::Normal))
@@ -284,7 +295,7 @@ fn build_dependencies_inner(
                 let id = &package.id;
                 // Return the name chosen in `Cargo.toml` and the path to the corresponding artifact
                 match artifacts.remove(id) {
-                    Some(Ok(artifacts)) => Some(Ok((name.replace('-', "_"), artifacts))),
+                    Some(Ok((_, artifacts))) => Some(Ok((name.replace('-', "_"), artifacts))),
                     Some(Err(what)) => Some(Err(Errored {
                         command: Command::new(what),
                         errors: vec![],
@@ -306,6 +317,28 @@ fn build_dependencies_inner(
             .collect::<Result<Vec<_>, Errored>>()?;
         let import_paths = import_paths.into_iter().collect();
         let import_libs = import_libs.into_iter().collect();
+
+        if info.build_std.is_some() {
+            let mut build_std_crates = HashSet::new();
+            build_std_crates.insert("core");
+            build_std_crates.insert("alloc");
+            build_std_crates.insert("proc_macro");
+            build_std_crates.insert("panic_unwind");
+            build_std_crates.insert("compiler_builtins");
+            build_std_crates.insert("std");
+            build_std_crates.insert("test");
+            build_std_crates.insert("panic_abort");
+
+            for (name, artifacts) in artifacts
+                .into_iter()
+                .filter_map(|(_, artifacts)| artifacts.ok())
+            {
+                if build_std_crates.remove(name.as_str()) {
+                    dependencies.push((format!("noprelude:{name}"), artifacts));
+                }
+            }
+        }
+
         return Ok(Dependencies {
             dependencies,
             import_paths,
@@ -329,6 +362,9 @@ pub struct DependencyBuilder {
     /// The command to run can be changed from `cargo` to any custom command to build the
     /// dependencies in `crate_manifest_path`.
     pub program: CommandBuilder,
+    /// Build with [`build-std`](https://doc.rust-lang.org/1.78.0/cargo/reference/unstable.html#build-std),
+    /// which requires the nightly toolchain. The [`String`] can contain the standard library crates to build.
+    pub build_std: Option<String>,
 }
 
 impl Default for DependencyBuilder {
@@ -336,6 +372,7 @@ impl Default for DependencyBuilder {
         Self {
             crate_manifest_path: PathBuf::from("Cargo.toml"),
             program: CommandBuilder::cargo(),
+            build_std: None,
         }
     }
 }
@@ -381,6 +418,11 @@ pub fn build_dependencies(
 ) -> Result<Vec<OsString>, Errored> {
     let dependencies = build_dependencies_inner(config, info)?;
     let mut args = vec![];
+
+    if info.build_std.is_some() {
+        args.push("-Zunstable-options".into());
+    }
+
     for (name, artifacts) in dependencies.dependencies {
         for dependency in artifacts {
             args.push("--extern".into());