+++ CHANGELOG.md
@@ -9,6 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Fixed
+
+### Changed
+
+### Removed
+
+
+## [0.24.0] - 2024-07-11
+
+### Added
+
 * `Text::diff()` creates a text status emitter that does not do full dumps of test stderr/stdout, but only emits the diff of the changes
 * Support `-Zbuild-std` by add
     * use `DependencyBuilder::build_std` to enable it